backend/pyproject.toml
[tool.poetry]
name = "centralisedbackend"
version = "0.1.0"
description = ""
authors = ["Akash Banger <akashbanger2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.3.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
